requirements.txt
fastapi
uvicorn
pydantic
langchain
llama-index
sentence-transformers
faiss-cpu
transformers
torch
pytesseract
opencv-python
pillow
python-dotenv
pdfplumber
tqdm
numpy
pandas
scikit-learn


